rename: MyActivity 폴더 이동
