+++ src/App.tsx
@@ -1,7 +1,7 @@
 import { createBrowserRouter, RouterProvider } from "react-router-dom";
 import DefaultLayout from "./layouts/DefaultLayout";
 import Board from "./pages/Board";
-import MyActivity from "./pages/MyActivity";
+import MyActivity from "./pages/my_activity/MyActivity";
 
 const router = createBrowserRouter([
   {
